requirements.txt
boto==2.38.0
csvkit==0.9.1
Flask==0.10.1
Flask-Bcrypt==0.6.2
Flask-Login==0.2.11
Flask-WTF==0.11
Flask-Cache
Jinja2==2.7.3
psycopg2==2.6
python-dateutil==2.2
pytz==2014.10
SQLAlchemy==0.9.9
